client/gtk2: Update capabilities if "retrieve-surrounding" is failed
Some engine developers wish to update the capabilities if emitting
"retrieve-surrounding" signal is failed so that engines disable the
surrounding text feature but I had deleted the logic with #2054 since
the second typing could be succeeded.
Asking all applications the second typing would not be a good condition
and the special issue should be fixed in firefox.

Fixes: https://github.com/ibus/ibus/commit/7b3b8c8b0c

BUG=https://github.com/ibus/ibus/issues/2054
BUG=https://github.com/ibus/ibus/issues/2354