bus: Avoid to hang the process with SIGUSR1 signal
Seems multiple `Xephyr -query` to a remote GDM login session causes
to leave the some processes of ibus-daemon even if the Xephyr sessions
are terminated and the issue is caused by calling bus_server_quit()
with SIGUSR1 signal.

I cannot reproduce the issue but calling the unsafe code from a signal
handler might lead to hangs and this is another signal connection
not to hang ibus-daemon.

BUG=https://github.com/ibus/ibus/pull/2499