github/hooks: Add post-merge
You can copy this file to .git/hooks/post-merge with the file permission
'0755'.
This script is useful not to commit invalid commit IDs every time and
leave a temporary version(1.5.00) in source files when a release is
bumped.
Unfortunately .git/hooks/pre-merge-commit does not run for git merge
fast-forwarding and I use post-merge instead as a workaround
but post-merge succeeds the last merge so need to run git reset --hard
if this script outputs warnings.