github/hooks: Add ibus-commit-common
You can copy .github/hooks/* file to .git/hooks/. with the file
permission '0755'.
This script is useful not to commit invalid commit IDs every time and
leave a temporary version(1.5.00) in source files when a release is
bumped.
Now a tiny spell checking is also added to output a warning of
possible typos but not an error.

Fixes: https://github.com/ibus/ibus/commit/3bbec69
Fixes: https://github.com/ibus/ibus/commit/cac49dc
Fixes: https://github.com/ibus/ibus/commit/5ed9f13
Fixes: https://github.com/ibus/ibus/commit/7c61e22