ui/gtk3: Set Indicator status at launching time
Panel.load_settings() is called before the "registered-status-notifier-item"
signal is emitted in Indicator and the Indicator's status was always
active at the launching time in the Plasma desktop session.
Now it's fixed.

Also the preferences menu in the activate menu is now shown in the
Plasma Xorg desktop session too.

BUG=https://github.com/ibus/ibus/issues/2463