ui/gtk3: Configure initial keymaps with localectl in Wayland
`setxkbmap -query` returns US layout with Xwayland and it does not
correspond to the session keymaps so ibus-panel now uses
`localectl status` in Wayland but it does not provide the session
XKB options against `setxkbmap` command in Xorg.
Need to think how to load or set the XKB options in Wayland later.

BUG=rhbz#2267615
BUG=https://github.com/ibus/ibus/pull/2624