src/ibusenginesimple: Do not update zero length preedit text
Several engines can inherit IBusEngieSimple for the compose key support
likes Anthy, Hangul, M17n and it could have the duplicated preedit text
between the engine and the parent IBusEngineSimple and it could update
the preedit text mutually by mistake.

Then the preedit text should not be hidden for zero length at least.
This update might not be enough but hope to fix the cursor position
reset with hiding the preedit text against the reported issue with
`m17n:sa:itrans` engine.

BUG=https://github.com/ibus/ibus/issues/2536