src/ibusengine: Hide preedit text with PanelExtensionReceived signal
IBusEngineSimple no longer calls to hide the preedit with the zero
length to fix the slurring cursor position but the emoji preedit
becomes not to be hidden with the preedit commit or escape events
as a side effect.
To fix the issue, IBusEngine needs to decide to clear or keep
the preedit when the emoji mode is finished.

BUG=https://github.com/ibus/ibus/issues/2536
Fixes: https://github.com/ibus/ibus/commit/a3a5a20a