bus: Change IBus unique name to :1.0 from IBUS_SERVICE_IBUS
According to the D-Bus specification,
`org.freedesktop.DBus.GetNameOwner()` must return a unique name (i.e.
one of the form `:1.23`). The only unique name which is allowed to not
be in this form is `org.freedesktop.DBus`.

The change in GLib 2.80.1 was to validate that the sender of every
message is either a unique name (according to `g_dbus_is_unique_name()`),
or the special name `org.freedesktop.DBus` to meet the specification.

In the long term, it might be more robust to rework the code so that the
mapping from `org.freedesktop.IBus` → a unique name is stored in
`dbus->names` (in `dbusimpl.c`) like all the other requested names.
However, handling for the `org.freedesktop.IBus` well-known name is
hardcoded throughout `dbusimpl.c`, so porting this single bit of it to
create a `BusNameService` would probably cause more problems.

BUG=https://github.com/ibus/ibus/issues/2639
BUG=https://gitlab.gnome.org/GNOME/glib/-/issues/3268
BUG=https://github.com/advisories/GHSA-f632-c3rh-r2v2