Fix Flatpak build for dnf and pyoverridesdir
`dnf` is not available during Flatpak builds with `flatpak-builder` and
the steps are omitted in autogen.sh to check FLATPAK_ID environment
variable.

When IBus is built with the Json files in test directory, the build
tries to install IBus.py file into pyoverridesdir likes
/usr/lib*/python3.12/site-packages/gi/overrides

But the directory is readonly with `flatpak-builder` and could be
failed.
Now `--disable-pygobject` option is added to `configure`.

IBus modules are embedded in Flatpak platforms likes GNOME, Freedesktop
and this issue might not happen in the platform builds since
the pyoverridesdir does not exist yet so this might fix the build
issues with Json files only in test directory.