src/tests: Fix test cases in GNOME Xorg in ibus-compose
"focus-in" signal is sent to IBus GtkIMModule with a delay in GNOME
desktop session and the signal order is not fixed with GTK_PHASE_TARGET
also "focus-in" signal is also sent with GtkText "realize"' signal.
Now the test cases wait for GtkText 'realize' signal.
This fixes the test cases in GNOME Xorg but there are still sone issues
in GNOME Wayland.

BUG=https://github.com/ibus/ibus/pull/2657