client/wayland: Copy system xkb_keymap for compose keys
The Wayland input-method can receive the system xkb_keymap
which is configured by the Wayland compositor.
I could not find a clear way to extract the compose options from
the system xkb_keymap or xkb_state and now refer the keyboard
to check if a key is assigned to the compose key.

Also now IBus refers XKB_DEFAULT_OPTIONS environment variable
to relative with the struct xkb_rule_names.options. [1]

[1] https://github.com/xkbcommon/libxkbcommon/blob/master/include/xkbcommon/xkbcommon.h#L393

BUG=https://github.com/ibus/ibus/issues/2579