ui/gtk3: Fix time lag of CandidatePanel in X11
The timed `set_lookup_table()` is a workaround with the Wayland
input-method protocol under the construction to stop the many D-Bus
methods but it should not affect Xorg UI at least.

Now `m_is_wayland flag` is inherited to the CandidatePanel and Switcher
and add `m_first_set_lookup_table` flag newly to enhance the lag
with the Wayland input-method protocol.

Also fix compile warnings with `valac` in `xkblayout.vala`.

Fixes: https://github.com/ibus/ibus/commit/d5e6e71
BUG=https://github.com/ibus/ibus/issues/2740