client/wayland: Fix infinite Return key in xterm
Currently the repeating key feature of IBus is disabled when the focus
is changed but the feature of the Wayland compositor still happens in
the Sway desktop session when invoke xterm by manual in foot.
I think Sway should stop the repeating key feature when the focus is
changed from the Wayland application to the non-Wayland application.

I add a workaround to call `zwp_virtual_keyboard_v1_key()` with
the `WL_KEYBOARD_KEY_STATE_RELEASED` flag to stop the feature of
the Sway compositor after the IBus repeating key timeout.