Add workaround until GNOME handle IBusMessage
The IBusEngine's focus-id feature is now always used in case the engine
is IBusEngineSimple to make a workaround in GNOME and Wayland because
GNOME-Shell does not implement the new D-Bus method
"SendMessageReceived" yet and the Wayland input-method protocol cannot
get the input surface for the Wayland xdg-system-bell protocol.

The auxiliary text is shown with the candidate popup in GNOME as a
workaround to notify the Unicode loading message until GNOME-Shell
implements the D-Bus method "SendMessageReceived".

Update the POT file.

BUG=https://github.com/ibus/ibus/issues/2722