client/wayland: Destroy virtual keyboard with IM deactivate
Labwc 0.8.3 deactivates the input-method with Alt-Tab key to
switch the activated windows and move the focus to the switcher popup.
The compositor does not release Alt-Tab key if IBus does not destroy
the virtual keyboard.
Now `zwp_virtual_keyboard_v1` is destroyed with
`input_method_deactivate_v2()` to release the pressed keys.

Also revert the commit 1f35eaa of "client/wayland: Fix infinite Return
key in xterm" which is no longer needed.

Fixes: https://github.com/ibus/ibus/commit/1f35eaa
BUG=https://github.com/ibus/ibus/issues/2742