bus/inputcontext: Fix reset signal w/ GTK_IM_MODULE=ibus in Wayland
Some applications send the reset signal when the focus is changed
in case that the IBus candidate popup takes the focus with non-Wayland
applications likes XIM and GTK2 in Wayland, and also GTK_IM_MODULE=ibus
with GTK3 and GTK4. The most IBus engines clear the lookup table with
the reset signal so ibus-daemon needs to ignore the reset signal
besides the focus-out signal for non-Wayland applications.

Also fix a memory leak in the D-Bus method.

Fixes: https://github.com/ibus/ibus/commit/e27acd3
Fixes: https://github.com/ibus/ibus/commit/1b67352
BUG=https://github.com/ibus/ibus/issues/2803