bus: Update gdm bits based on new gdm changes
Since GDM 49 version, gdm user doesn't exist and for each greeter
there's a dynamic user creation, e.g. gdm-greeter-1, gdm-greeter-2.
All of them are part of gdm group, so instead of checking for a specific
user, we check for a specific group.

Also, remove Conflicts in systemd service to allow ibus in GDM sessions.
This covers some cases where users might have special characters in their
password or username.

The original patch was reverted by the patch(ea70b2d) to disable IBus in
GDM at once since the change would a big one for the code freeze term.
However, we decided to bring back the patch to fix the compose keys of
the Portuguese keyboard because GDM had enabled the compose keys only
with IBus in fact. It would not be good to keep showing the compose
pre-edit in the GDM password entry, and I will consider better
suggestions for that later.

Fixes: https://github.com/ibus/ibus/commit/ea70b2d
BUG=https://gitlab.gnome.org/GNOME/gdm/-/issues/1018
BUG=https://github.com/ibus/ibus/pull/2804